+++ packages/core/src/index.ts
@@ -12,5 +12,6 @@ export * from "./notifications.js";
 export * from "./referrals.js";
 export * from "./vip.js";
 export * from "./staking.js";
+export * from "./launchpad.js";
 export * from "./wallet/derivation.js";
 export * from "./wallet/provision.js";